backend/requirements.txt
flask
pymongo
python-dotenv
werkzeug
flask-jwt-extended
flask-cors